refactor: move usePermissions
